# Memory footprint gc latency level