--- a/src/gc/gccommon.cpp
+++ b/src/gc/gccommon.cpp
@@ -53,8 +53,6 @@ void record_global_mechanism (int mech_index)
 }
 #endif //GC_CONFIG_DRIVEN
 
-int32_t g_bLowMemoryFromHost = 0;
-
 #ifdef WRITE_BARRIER_CHECK
 
 #define INVALIDGCVALUE (void *)((size_t)0xcccccccd)