--- a/src/gc/gcconfig.h
+++ b/src/gc/gcconfig.h
@@ -83,7 +83,7 @@ public:
   INT_CONFIG(LatencyMode,   "GCLatencyMode", -1,                                               \
       "Specifies the GC latency mode - batch, interactive or low latency (note that the same " \
       "thing can be specified via API which is the supported way")                             \
-  INT_CONFIG(LatencyLevel,  "GCLatencyLevel", 2,                                               \
+  INT_CONFIG(LatencyLevel,  "GCLatencyLevel", 1,                                               \
       "Specifies the GC latency level that you want to optimize for. Must be a number from 0"  \
       "3. See documentation for more details on each level.")                                  \
   INT_CONFIG(LogFileSize,   "GCLogFileSize", 0, "Specifies the GC log file size")              \