@gbalykov 2017-12-12
@dotnet-bot test Tizen armel Cross Checked Innerloop Build and Test
@dotnet-bot test Windows_NT x64 Innerloop Formatting