@gbalykov 2018-01-17
@Maoni0 @swgillespie

We have measured CoreFX performance tests on Windows x64 on this change and results are the next: geomean of "memory_consumption" level (relative to base) is 1.01381 (-1.381%). This includes only the tests with stdev less than 1%.

What do you think about this change and the results?

cc @kvochko @Dmitri-Botcharnikov